refactor: moved some directories
