+++ src/app/page.tsx
@@ -1,10 +1,8 @@
-"use client";
 import Hero from "@/block/Home/Hero/Hero";
 import Presentation from "@/block/Home/Presentation/Presentation";
 import CampaignDemo from "@/block/Home/CampaignDemo/CampaignDemo";
 import HomeWorldDemo from "@/block/Home/WorldDemo/WorldDemo";
 import UsecaseDemo from "@/block/Home/UsecaseDemo/UsecaseDemo";
-import {useSession} from "next-auth/react";
 
 const Home = () => {
 
@@ -17,7 +15,7 @@ const Home = () => {
             <UsecaseDemo/>
         </main>
     );
-    
+
 };
 
 export default Home;
