+++ src/block/index.ts
@@ -1,6 +1,7 @@
 import {Header, Footer, Hero, Presentation, CampaignDemo, UsecaseDemo, WorldDemo} from "@/block/Home";
 import Campaigns from "@/block/Campaigns/Campaigns";
-import Profile from "@/block/Profile/Profile";
+import ProfileInfo from "@/block/Profile/ProfileInfo/ProfileInfo";
+import ProfileBanner from "@/block/Profile/ProfileBanner/ProfileBanner";
 import SectionCharacters from "@/block/CampaignCreation/SectionCharacters/SectionCharacters";
 import SectionInfos from "@/block/CampaignCreation/SectionInfos/SectionInfos";
 import SectionMap from "@/block/CampaignCreation/SectionMap/SectionMap";
@@ -8,6 +9,6 @@ import SectionQuest from "@/block/CampaignCreation/SectionQuest/SectionQuest";
 
 export {
     Header, Footer, Hero, Presentation, CampaignDemo, UsecaseDemo, WorldDemo,
-    Campaigns, Profile,
+    Campaigns, ProfileInfo, ProfileBanner,
     SectionCharacters, SectionInfos, SectionMap, SectionQuest
 };
